+++ packages/ui/src/components/common/types.ts
@@ -1,0 +1,7 @@
+export interface OnePagerCommentItem {
+  id: string;
+  author: string;
+  content: string;
+  profileImage?: string | null;
+  type?: "GENERAL" | "REJECTION_REASON";
+}
